Change postcode lookup tests to reflect new content
Have changed the assertion in the
'then_i_can_see_the_postcode_lookup_form' method to look for the
'royal_mail_lookup.link_text' due to difficulties in making assertions
on yaml that is sanitized. It should provide an acceptable stand-in for
the 'input_label' given they are both contained to the 'postcode-lookup'
div.